minor typing
